+++ PlateUp-reactnative/constants/utils.js
@@ -1,3 +1,17 @@
+import {
+  AlertIOS, Platform, ToastAndroid
+} from 'react-native';
+
+export const iPhoneX = () => Platform.OS === 'ios' && (height === 812 || width === 812);
+
+export function toast(msg) {
+  if (Platform.OS === 'android') {
+    ToastAndroid.show(msg, ToastAndroid.LONG);
+  } else {
+    AlertIOS.alert(msg);
+  }
+}
+
 export function constructQueryParams(params, keyTransforms) {
   let queryParams = "";
 
@@ -13,9 +27,9 @@ export function constructQueryParams(params, keyTransforms) {
 
 function appendQueryParam(existingQueryParams, newParam) {
   if (existingQueryParams) {
-    existingQueryParams += (`&${newParam}`);
+    existingQueryParams += `&${newParam}`;
   } else {
-    existingQueryParams = (`?${newParam}`);
+    existingQueryParams = `?${newParam}`;
   }
 
   return existingQueryParams;
